requirements.txt
Django==5.1.2
djangorestframework
djangorestframework-simplejwt
